fix(plugins): move live-modules project config into plugins/api
Files plugin must not import host renderer modules; share the config
load/cache with Settings via plugins/api (same bus pattern as
files-disk-path-opened).
