+++ src/plugins/api/live-modules-project-config.ts
