+++ src/plugins/builtin/files/renderer/preview/load-live-modules-config.ts
@@ -1,6 +1,7 @@
 /**
  * Files-panel entry for Live Modules project config.
- * Implementation lives under renderer/lib (shared with Project → General).
+ * Implementation lives under plugins/api (shared with Project → General;
+ * plugins must not import src/renderer host modules).
  */
 export {
   applyLiveModulesProjectConfigAfterSave,
@@ -10,4 +11,4 @@ export {
   notifyLiveModulesProjectConfigChanged,
   resetLiveModulesProjectConfigCacheForTests,
   subscribeLiveModulesProjectConfigChanged,
-} from "@/lib/live-modules/project-config-cache.ts";
+} from "@plugins/api/live-modules-project-config-cache.ts";
